refactor for better compatibility with ee
